Initial commit for migration from create react app to vite.
