+++ package.json
@@ -1,12 +1,15 @@
 {
   "name": "pledger-ui",
   "version": "1.4.0",
+  "type": "module",
   "private": true,
   "proxy": "http://127.0.0.1:8080/",
   "homepage": "/ui",
   "dependencies": {
     "@mdi/js": "^7.4.47",
     "@mdi/react": "^1.6.1",
+    "@tailwindcss/vite": "^4.0.7",
+    "@vitejs/plugin-react-swc": "^3.8.0",
     "axios": "^1.7.2",
     "chart.js": "^4.4.7",
     "chartjs-adapter-spacetime": "^1.0.1",
@@ -18,8 +21,10 @@
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.53.1",
     "react-router": "^7.1.1",
-    "react-scripts": "5.0.1",
-    "spacetime": "^7.6.2"
+    "spacetime": "^7.6.2",
+    "vite": "^6.1.1",
+    "vite-plugin-svgr": "^4.3.0",
+    "vite-tsconfig-paths": "^5.1.4"
   },
   "devDependencies": {
     "@stylistic/eslint-plugin": "^4.0.0",
@@ -29,22 +34,26 @@
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.14",
     "@types/react": "^19.0.2",
+    "@vitest/coverage-v8": "^3.0.6",
     "eslint": "^9.14.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "jsdom": "^26.0.0",
     "prop-types": "^15.8.1",
     "sass": "^1.83.0",
+    "sass-embedded": "^1.85.0",
     "tailwindcss": "^3.4.14",
     "ts-jest": "^29.2.5",
     "typescript": "^5.6.3",
     "typescript-eslint": "8.24.1",
+    "vitest": "^3.0.6",
     "web-vitals": "^4.2.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "tsc && vite build",
+    "test": "vitest",
+    "test:coverage": "vitest run --coverage --watch=false",
     "lint": "eslint src/**/*.ts{,x}",
     "lint:fix": "eslint src/**/*.ts{,x} --fix"
   },
