+++ src/assets/css/BreadCrumbs.scss
@@ -1,4 +1,4 @@
-@import "../../config/constants";
+@import "constants";
 
 .hamburger-menu {
   display: none;
