+++ src/assets/css/Sidebar.scss
@@ -1,4 +1,4 @@
-@import 'src/config/constants';
+@import 'constants';
 
 $footer-height: 45px;
