--- src/config/constants.scss
@@ -1,12 +1,0 @@
-$media-max-width: 991.98px;
-
-$tabled-portrait-max-width: 1024px;
-
-$phone-landscape-max-width: 768px;
-$phone-portrait-max-width: 420px;
-
-$sidebar-offset: 218px;
-$sidebar-min-offset: 50px;
-$top-bar-height: 55px;
-
-$transition-time: .5s;
