+++ vite.config.mts
@@ -1,0 +1,32 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react-swc'
+import tailwindcss from '@tailwindcss/vite'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+    base: '/ui',
+    plugins: [
+        react(),
+        tailwindcss()
+    ],
+    server: {
+        proxy: {
+            '/api': {
+                target: 'http://finance.local',
+                changeOrigin: true
+            }
+        }
+    },
+    test: {
+        globals: true,
+        environment: 'jsdom',
+        setupFiles: './src/setupTests.js',
+        css: true,
+        reporters: ['verbose'],
+        coverage: {
+            reporter: ['text', 'json', 'html'],
+            include: ['src/**/*'],
+            exclude: [],
+        }
+    },
+})
